requirements.txt
torch
transformers
peft
fastapi
uvicorn
python-multipart
Pillow
pandas
